Put boilerplate code into MetadataTypeSystemContext
There is a bunch of boilerplate code every user of the type system needs
to copy to actually use the type system. The actions are specific to
metadata based worlds - which means they fundamentally can't be in
TypeSystemContext. But they can be in a MetadataTypeSystemContext.

I'm also adding a throwing implementation for the various algorithm
retrieval methods. There are in fact uses of the type system (outside of
tests) that work just fine without providing an implementation. We
shouldn't force type system users to provide one.
