+++ src/ILCompiler.MetadataTransform/tests/TestTypeSystemContext.cs
@@ -14,65 +14,9 @@ using System.IO;
 
 namespace MetadataTransformTests
 {
-    class TestTypeSystemContext : TypeSystemContext
+    class TestTypeSystemContext : MetadataTypeSystemContext
     {
-        static readonly string[] s_wellKnownTypeNames = new string[] {
-            "Void",
-            "Boolean",
-            "Char",
-            "SByte",
-            "Byte",
-            "Int16",
-            "UInt16",
-            "Int32",
-            "UInt32",
-            "Int64",
-            "UInt64",
-            "IntPtr",
-            "UIntPtr",
-            "Single",
-            "Double",
-
-            "ValueType",
-            "Enum",
-            "Nullable`1",
-
-            "Object",
-            "String",
-            "Array",
-            "MulticastDelegate",
-
-            "RuntimeTypeHandle",
-            "RuntimeMethodHandle",
-            "RuntimeFieldHandle",
-        };
-
-        MetadataType[] _wellKnownTypes = new MetadataType[s_wellKnownTypeNames.Length];
-
-        EcmaModule _systemModule;
-
         Dictionary<string, EcmaModule> _modules = new Dictionary<string, EcmaModule>(StringComparer.OrdinalIgnoreCase);
-
-        public override DefType GetWellKnownType(WellKnownType wellKnownType)
-        {
-            return _wellKnownTypes[(int)wellKnownType - 1];
-        }
-
-        public void SetSystemModule(EcmaModule systemModule)
-        {
-            _systemModule = systemModule;
-
-            // Sanity check the name table
-            Debug.Assert(s_wellKnownTypeNames[(int)WellKnownType.MulticastDelegate - 1] == "MulticastDelegate");
-
-            // Initialize all well known types - it will save us from checking the name for each loaded type
-            for (int typeIndex = 0; typeIndex < _wellKnownTypes.Length; typeIndex++)
-            {
-                MetadataType type = _systemModule.GetType("System", s_wellKnownTypeNames[typeIndex]);
-                type.SetWellKnownType((WellKnownType)(typeIndex + 1));
-                _wellKnownTypes[typeIndex] = type;
-            }
-        }
 
         public EcmaModule GetModuleForSimpleName(string simpleName)
         {
@@ -93,21 +37,6 @@ namespace MetadataTransformTests
         public override ModuleDesc ResolveAssembly(System.Reflection.AssemblyName name, bool throwIfNotFound)
         {
             return GetModuleForSimpleName(name.Name);
-        }
-
-        public override FieldLayoutAlgorithm GetLayoutAlgorithmForType(DefType type)
-        {
-            throw new NotImplementedException();
-        }
-
-        public override RuntimeInterfacesAlgorithm GetRuntimeInterfacesAlgorithmForNonPointerArrayType(ArrayType type)
-        {
-            throw new NotImplementedException();
-        }
-
-        public override RuntimeInterfacesAlgorithm GetRuntimeInterfacesAlgorithmForMetadataType(MetadataType type)
-        {
-            throw new NotImplementedException();
         }
     }
 }
